requirements.txt
asgiref==3.11.0
Django==5.2.9
django-allauth==65.13.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
